call sessionListener when requestiong session
request session takes a callback which sets the session locally, however requesting a session never calls the callbacks set in sessionListeners which maintain the state of our session info.  It is therefore necessary to call sessionListener here explicitly in order to make sure the housekeeping gets done when requesting a session.